cron connected to changed endpoints in own file now
